Added generic types
One problem I've been trying to solve is, how do I design the Store's
state and reducer actions to be flexible enough to include whatever
data a user of this library could possibly want to include in their
battles, but avoid having the library be type unsafe or clumsy to use?

I've decided that my best bet is to make the Store's State and Action
classes emprt and abstract, forcing users of the library to write
those classes themselves, then use generics extensively to ensure that
all of the classes that interact with the Store will be forced to use
consumers' custom Actions and States correctly at compile-time. This
gives consumers the flexibility to put absolutely whatever data they
wish into the State and Actions, but also guarantees that other
classes will know exactly how to query the State and Actions to get
the data they contain, and the compiler and programming tools can help
programmers know that they're using their State and Action classes
correctly as well.

The downside, of course, is that now the library has a ton of
generics all over the place, making the library more complicated to
use. I think this is acceptable: It increases the library's learning
curve, but ensures that once someone has understood how to use the
library, it will be much more difficult to misuse it. This puts the
onus on me to write good documentation for the library with examples
that help users understand how to use it at a basic level more
quickly!

Once I settled on using generics for the State and Action types, I saw
several other opportunities to use the same technique elsewhere. I
will continue to use the same technique as I continue to design this
library: Whenever I want the library's user to decide what data needs
to go in a class, I will force the user to write the class themselves
but use generics to make sure my classes coordinate on using the
custom classes correctly.